moved User class into user.py
